introduced int_t for various reasons; not all interfaces/tests etc. are guaranteed to work, careful checks needed; updated octave interface; tried to reflect changes due to SQProblemSchur class in C and Python interface, more tests needed

+++ src/BLASReplacement.cpp
@@ -35,12 +35,13 @@
 #include <qpOASES/Utils.hpp>
 
 
-extern "C" void dgemm_ ( const char *TRANSA, const char *TRANSB,
-		const unsigned long *M, const unsigned long *N, const unsigned long *K,
-		const double *ALPHA, const double *A, const unsigned long *LDA, const double *B, const unsigned long *LDB,
-		const double *BETA, double *C, const unsigned long *LDC)
+extern "C" void dgemm_(	const char *TRANSA, const char *TRANSB,
+						const unsigned long *M, const unsigned long *N, const unsigned long *K,
+						const double *ALPHA, const double *A, const unsigned long *LDA, const double *B, const unsigned long *LDB,
+						const double *BETA, double *C, const unsigned long *LDC
+						)
 {
-	unsigned int i, j, k;
+	unsigned long i, j, k;
 
 	if ( REFER_NAMESPACE_QPOASES isZero(*BETA) == REFER_NAMESPACE_QPOASES BT_TRUE )
 		for (k = 0; k < *N; k++)
@@ -89,12 +90,13 @@ extern "C" void dgemm_ ( const char *TRANSA, const char *TRANSB,
 						C[j+(*LDC)*k] += *ALPHA * A[i+(*LDA)*j] * B[i+(*LDB)*k];
 }
 
-extern "C" void sgemm_ ( const char *TRANSA, const char *TRANSB,
-		const unsigned long *M, const unsigned long *N, const unsigned long *K,
-		const float *ALPHA, const float *A, const unsigned long *LDA, const float *B, const unsigned long *LDB,
-		const float *BETA, float *C, const unsigned long *LDC)
+extern "C" void sgemm_(	const char *TRANSA, const char *TRANSB,
+						const unsigned long *M, const unsigned long *N, const unsigned long *K,
+						const float *ALPHA, const float *A, const unsigned long *LDA, const float *B, const unsigned long *LDB,
+						const float *BETA, float *C, const unsigned long *LDC
+						)
 {
-	unsigned int i, j, k;
+	unsigned long i, j, k;
 
 	if ( REFER_NAMESPACE_QPOASES isZero(*BETA) == REFER_NAMESPACE_QPOASES BT_TRUE )
 		for (k = 0; k < *N; k++)
